Moved get_files function to utilities
