+++ source/utils.py
@@ -152,10 +152,18 @@ class WriteTextMidiToFile:  # utils saving to file
         return self.output_path_filename
 
 
-def get_files(directory, extension):
-    """Given a directory, get a list of the file paths of all files matching the
-    specified file extension."""
-    return directory.glob(f"*.{extension}")
+def get_files(directory, extension, recursive=False):
+    """
+    Given a directory, get a list of the file paths of all files matching the
+    specified file extension.
+    directory: the directory to search as a Path object
+    extension: the file extension to match as a string
+    recursive: whether to search recursively in the directory or not
+    """
+    if recursive:
+        return list(directory.rglob(f"*.{extension}"))
+    else:
+        return list(directory.glob(f"*.{extension}"))
 
 
 def timeit(func):
